+++ packages/chains/chains/333000333.ts
@@ -5,7 +5,7 @@ export default {
   "chain": "MELD",
   "rpc": [
     "https://meld.rpc.thirdweb.com/${THIRDWEB_API_KEY}",
-    "https://network-rpc.meld.com"
+    "https://subnets.avax.network/meld/mainnet/rpc"
   ],
   "faucets": [],
   "features": [],
@@ -24,7 +24,19 @@ export default {
   "shortName": "meld",
   "chainId": 333000333,
   "networkId": 333000333,
-  "explorers": [],
+  "explorers": [
+    {
+      "name": "explorer",
+      "url": "https://subnets.avax.network/meld",
+      "icon": {
+        "url": "ipfs://QmRhB4AbjDrhvwfSAQi2JvKirFiDWxzJvKEvG8S8AdDdED",
+        "width": 4000,
+        "height": 4000,
+        "format": "png"
+      },
+      "standard": "EIP3091"
+    }
+  ],
   "testnet": false,
   "slug": "meld"
 } as const satisfies Chain;